权限管理:角色权限CRUD,订单详情(改)

+++ src/main/java/com/shsxt/crm/controller/OrderDetailController.java
@@ -25,7 +25,7 @@ public class OrderDetailController extends BaseController<OrderDetail> {
     @RequestMapping("getTotalPrice")
     @ResponseBody
     public ResultInfo getTotalPrice(Integer orderId){
-        Long total=orderDetailService.getTotalPrice(orderId);
+        Float total=orderDetailService.getTotalPrice(orderId);
         return  success(total);
     }
